minor restructuring of database package
