+++ src/main/scala/com/campudus/tableaux/controller/DemoController.scala
@@ -7,6 +7,7 @@ package com.campudus.tableaux.controller
 import com.campudus.tableaux.ArgumentChecker._
 import com.campudus.tableaux.database.Tableaux._
 import com.campudus.tableaux.database._
+import com.campudus.tableaux.database.structure.CreateColumn
 import com.campudus.tableaux.helper.HelperFunctions._
 import com.campudus.tableaux.helper.StandardVerticle
 import org.vertx.scala.core.json.Json
