+++ src/main/scala/com/campudus/tableaux/database/structure/returntype.scala
@@ -1,0 +1,9 @@
+package com.campudus.tableaux.database.structure
+
+sealed trait ReturnType
+
+case object GetReturn extends ReturnType
+
+case object SetReturn extends ReturnType
+
+case object EmptyReturn extends ReturnType
