+++ src/main/scala/com/campudus/tableaux/database/structure/row.scala
@@ -1,0 +1,23 @@
+package com.campudus.tableaux.database.structure
+
+import com.campudus.tableaux.database.DomainObject
+import com.campudus.tableaux.database.Tableaux._
+import org.vertx.scala.core.json._
+
+case class RowIdentifier(table: Table, id: IdType) extends DomainObject {
+  def getJson: JsonObject = Json.obj("rows" -> Json.arr(Json.obj("id" -> id)))
+
+  def setJson: JsonObject = Json.obj("rows" -> Json.arr(Json.obj("id" -> id)))
+}
+
+case class Row(table: Table, id: IdType, values: Seq[_]) extends DomainObject {
+  def getJson: JsonObject = Json.obj("rows" -> Json.arr(Json.obj("id" -> id, "values" -> values)))
+
+  def setJson: JsonObject = Json.obj("rows" -> Json.arr(Json.obj("id" -> id)))
+}
+
+case class RowSeq(rows: Seq[Row]) extends DomainObject {
+  def getJson: JsonObject = Json.obj("rows" -> (rows map { r => Json.obj("id" -> r.id, "values" -> r.values) }))
+
+  def setJson: JsonObject = Json.obj("rows" -> (rows map { r => Json.obj("id" -> r.id) }))
+}
