+++ src/main/scala/com/campudus/tableaux/helper/HelperFunctions.scala
@@ -3,6 +3,7 @@ package com.campudus.tableaux.helper
 import com.campudus.tableaux.ArgumentChecker._
 import com.campudus.tableaux.database.Tableaux.{IdType, LinkConnection}
 import com.campudus.tableaux.database._
+import com.campudus.tableaux.database.structure.{Mapper, TableauxDbType, LinkType, CreateColumn}
 import com.campudus.tableaux.{ArgumentCheck, FailArg, InvalidJsonException, OkArg}
 import org.vertx.scala.core.json.{JsonArray, JsonObject}
 import com.campudus.tableaux.database.Tableaux.Ordering
