+++ src/main/scala/com/campudus/tableaux/router/BaseRouter.scala
@@ -1,7 +1,8 @@
 package com.campudus.tableaux.router
 
 import com.campudus.tableaux._
-import com.campudus.tableaux.database.{DomainObject, ReturnType}
+import com.campudus.tableaux.database.structure.ReturnType
+import com.campudus.tableaux.database.DomainObject
 import com.campudus.tableaux.helper.StandardVerticle
 import org.vertx.scala.core.http.HttpServerRequest
 import org.vertx.scala.core.json._
