+++ src/main/scala/com/campudus/tableaux/router/DemoRouter.scala
@@ -1,15 +1,10 @@
 package com.campudus.tableaux.router
 
-import com.campudus.tableaux.helper.HelperFunctions
-import HelperFunctions._
-import com.campudus.tableaux.controller.{DemoController, TableauxController}
-import com.campudus.tableaux.database.{DatabaseAccess, DatabaseConnection, SetReturn}
-import org.vertx.scala.core.Vertx
+import com.campudus.tableaux.controller.DemoController
+import com.campudus.tableaux.database.structure.SetReturn
+import com.campudus.tableaux.database.{DatabaseAccess, DatabaseConnection}
 import org.vertx.scala.core.http.HttpServerRequest
-import org.vertx.scala.core.json.Json
-import org.vertx.scala.core.logging.Logger
-import org.vertx.scala.platform.{Container, Verticle}
-import org.vertx.scala.router.Router
+import org.vertx.scala.platform.Verticle
 import org.vertx.scala.router.routing.Post
 
 class DemoRouter(val verticle: Verticle, val database: DatabaseConnection) extends BaseRouter with DatabaseAccess {
